+++ libs/librrgraph/src/base/rr_graph_storage.cpp
@@ -608,10 +608,10 @@ bool t_rr_graph_storage::validate(const vtr::vector<RRSwitchId, t_rr_switch_inf>
 }
 
 const char* t_rr_graph_storage::node_type_string(RRNodeId id) const {
-    return rr_node_typename[(size_t)node_type(id)];
+    return rr_node_typename[node_type(id)];
 }
 const char* t_rr_graph_view::node_type_string(RRNodeId id) const {
-    return rr_node_typename[(size_t)node_type(id)];
+    return rr_node_typename[node_type(id)];
 }
 
 const std::string& t_rr_graph_storage::node_direction_string(RRNodeId id) const {
@@ -675,7 +675,7 @@ static int get_node_pin_num(
     RRNodeId id) {
     e_rr_type node_type = node_storage[id].type_;
     if (node_type != e_rr_type::IPIN && node_type != e_rr_type::OPIN) {
-        VTR_LOG_ERROR("Attempted to access RR node 'pin_num' for non-IPIN/OPIN type '%s'", rr_node_typename[(size_t)node_type]);
+        VTR_LOG_ERROR("Attempted to access RR node 'pin_num' for non-IPIN/OPIN type '%s'", rr_node_typename[node_type]);
     }
     return node_ptc[id].ptc_.pin_num;
 }
@@ -686,7 +686,7 @@ static int get_node_track_num(
     RRNodeId id) {
     e_rr_type node_type = node_storage[id].type_;
     if (node_type != e_rr_type::CHANX && node_type != e_rr_type::CHANY) {
-        VTR_LOG_ERROR("Attempted to access RR node 'track_num' for non-CHANX/CHANY type '%s'", rr_node_typename[(size_t)node_type]);
+        VTR_LOG_ERROR("Attempted to access RR node 'track_num' for non-CHANX/CHANY type '%s'", rr_node_typename[node_type]);
     }
     return node_ptc[id].ptc_.track_num;
 }
@@ -697,7 +697,7 @@ static int get_node_class_num(
     RRNodeId id) {
     e_rr_type node_type = node_storage[id].type_;
     if (node_type != e_rr_type::SOURCE && node_type != e_rr_type::SINK) {
-        VTR_LOG_ERROR("Attempted to access RR node 'class_num' for non-SOURCE/SINK type '%s'", rr_node_typename[(size_t)node_type]);
+        VTR_LOG_ERROR("Attempted to access RR node 'class_num' for non-SOURCE/SINK type '%s'", rr_node_typename[node_type]);
     }
     return node_ptc[id].ptc_.class_num;
 }
